requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
albumentations==1.3.1
arrow==1.2.3
basicsr==1.4.2
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.25.0
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
diffusers==0.19.3
easydict==1.10
facexlib==0.3.0
filelock==3.12.2
insightface
filterpy==1.4.5
Flask==2.3.2
Flask-AutoIndex==0.6.6
Flask-Silk==0.2
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
fsspec==2023.6.0
future==0.18.3
gdown==4.7.1
gfpgan==1.3.8
glob2==0.7
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.4
kiwisolver==1.4.4
lazy_loader==0.3
lit==15.0.7
llvmlite==0.40.1
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
onnx==1.14.0
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==10.0.0
pip-autoremove==0.10.0
platformdirs==3.10.0
prettytable==3.8.0
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pyre-extensions==0.0.29
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
sympy==1.11.1
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
Werkzeug==2.3.6
xformers==0.0.20
yapf==0.40.1
zipp==3.16.2
